+++ src/App.js
@@ -1,27 +1,17 @@
-import Navigation from "./components/Nabvar";
-import LeftSidebar from "./components/LeftSidebar";
-import Content from "./components/Content";
-import RightSidebar from "./components/RightSidebar";
-import { useState } from "react";
+// import { useState } from "react";
 import "../src/App.scss";
 
+import { Routes, Route } from "react-router-dom";
+
+//containers
+import HomeContainer from "./containers/home/home";
 function App() {
-  const [mobileMenu, setmobileMenu] = useState(false);
-
-  const toggle = () => {
-    setmobileMenu(!mobileMenu);
-  };
-
   return (
-    <>
-      <Navigation openMenu={toggle} />
-
-      <main className="main-container">
-        <LeftSidebar burgerMenu={mobileMenu} closeMenu={toggle} />
-        <Content />
-        <RightSidebar />
-      </main>
-    </>
+    <div className="App">
+      <Routes>
+        <Route path="/" element={<HomeContainer />} />
+      </Routes>
+    </div>
   );
 }
